Support of Smartwares SHS-53000 radiator valve
With these version you can display a currentTemperature for the
thermostat if you have an external thermometer available, especially in
the same room.
In this case you add a configuration parameter « valves » in the
configuration file with one hash table mapping the correspondance.
For exemple :
"valves": {"ValveX":"ThermometerForX", "ValveY":"ThermometerForY"}
If you have no thermometers available, the current temperature display
the target temperature.